// src/codepage/tvm-spec.d.ts
export type VariableName = string;
/**
 * Hint for converting operands between raw values and Asm.fif display format
 */
export type DisplayHint =
  | {
      type: "continuation";
    }
  | {
      type: "dictionary";
      size_var: VariableName;
    }
  | {
      type: "add";
      value: number;
    }
  | {
      type: "stack";
    }
  | {
      type: "register";
    }
  | {
      type: "pushint4";
    }
  | {
      type: "optional_nargs";
    }
  | {
      type: "plduz";
    };